[cycles]: record m-cycles of each instruction.

+++ include/cpu.h
@@ -2,7 +2,7 @@
 #define CPU_H
 
 int cpu_init();
-int cpu_step();
+uint8_t cpu_step();
 
 uint8_t cpu_read_ie();
 void cpu_write_ie(uint8_t value);
